requirements.txt
# Pilar2 Project Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.0

# PDF Processing
reportlab>=4.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0

# XML Processing
lxml>=4.9.0
xmltodict>=0.13.0

# AI/ML (optional - comment out if causing issues)
# scikit-learn>=1.3.0
# transformers>=4.35.0
# torch>=2.1.0
nltk>=3.8.0

# OpenAI Integration
openai>=1.3.0

# Web Search Integration
serper-dev>=0.1.0

# Web Scraping Integration
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
lxml>=4.9.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# API & HTTP
requests>=2.31.0
httpx>=0.25.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.11.0
flake8>=6.1.0
# mypy>=1.7.0  # Commented out as it can cause issues

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.0
